+++ src/lib/index.ts
@@ -4,5 +4,5 @@ export { default as Handle } from '$lib/components/Handle.svelte';
 export { default as Remove } from '$lib/components/Remove.svelte';
 export { default as IconHandle } from '$lib/components/Icons/IconHandle.svelte';
 export { default as IconRemove } from '$lib/components/Icons/IconRemove.svelte';
-export { dispatch, removeItem, sortItems } from '$lib/utils/index.js';
+export { removeItem, sortItems } from '$lib/utils/index.js';
 export type { RemoveEventDetail, SortEventDetail, SortableItemData } from '$lib/types/index.js';