+++ src/lib/utils/index.ts
@@ -94,19 +94,14 @@ export function getCollidingItem(
 	return collidingItems[0];
 }
 
-export function dispatch(target: HTMLElement, name: string, detail: object) {
-	const event = new CustomEvent(name, { bubbles: true, detail });
-	target.dispatchEvent(event);
-}
-
 export function sortItems<T>(items: T[], from: number, to: number) {
-	const clone = structuredClone(items);
+	const clone = [...items];
 	clone.splice(to < 0 ? clone.length + to : to, 0, clone.splice(from, 1)[0]);
 	return clone;
 }
 
 export function removeItem<T>(items: T[], index: SortableItemProps['index']) {
-	const clone = structuredClone(items);
+	const clone = [...items];
 	clone.splice(index, 1);
 	return clone;
 }