+++ src/routes/+layout.svelte
@@ -1,16 +1,18 @@
 <script lang="ts">
 	import { version } from '$app/environment';
-	import { page } from '$app/stores';
-	import { props } from './stores.js';
+	import { page } from '$app/state';
+	import { rootProps } from './stores.js';
 	import './styles.css';
 
-	let isControlsExpanded = true;
+	let { children } = $props();
+
+	let isControlsExpanded = $state(true);
 </script>
 
 <div id="app" class="app" class:is-controls-expanded={isControlsExpanded}>
 	<main class="app-main">
 		<div class="container">
-			<slot />
+			{@render children?.()}
 		</div>
 	</main>
 	<aside id="app-controls" class="app-controls">
@@ -18,7 +20,7 @@
 			class="app-controls__toggle button"
 			aria-expanded={isControlsExpanded}
 			aria-controls="app-controls"
-			on:click={() => (isControlsExpanded = !isControlsExpanded)}
+			onclick={() => (isControlsExpanded = !isControlsExpanded)}
 		>
 			{isControlsExpanded ? 'Hide controls' : 'Show controls'}
 		</button>
@@ -36,14 +38,14 @@
 					<tr>
 						<td></td>
 						<td><label for="gap">gap</label></td>
-						<td><input id="gap" type="number" min="0" bind:value={$props.gap} /></td>
+						<td><input id="gap" type="number" min="0" bind:value={$rootProps.gap} /></td>
 						<td></td>
 					</tr>
 					<tr>
 						<td></td>
 						<td><label for="direction">direction</label></td>
 						<td>
-							<select id="direction" bind:value={$props.direction}>
+							<select id="direction" bind:value={$rootProps.direction}>
 								<option value="vertical">vertical</option>
 								<option value="horizontal">horizontal</option>
 							</select>
@@ -60,7 +62,7 @@
 								min="0.5"
 								max="2"
 								step="0.1"
-								bind:value={$props.swapThreshold}
+								bind:value={$rootProps.swapThreshold}
 							/>
 						</td>
 						<td></td>
@@ -73,7 +75,7 @@
 								id="transition-duration"
 								type="number"
 								min="0"
-								bind:value={$props.transitionDuration}
+								bind:value={$rootProps.transitionDuration}
 							/>
 						</td>
 						<td></td>
@@ -82,7 +84,7 @@
 						<td></td>
 						<td><label for="has-drop-marker">hasDropMarker</label></td>
 						<td>
-							<input id="has-drop-marker" type="checkbox" bind:checked={$props.hasDropMarker} />
+							<input id="has-drop-marker" type="checkbox" bind:checked={$rootProps.hasDropMarker} />
 						</td>
 						<td></td>
 					</tr>
@@ -90,7 +92,7 @@
 						<td></td>
 						<td><label for="has-locked-axis">hasLockedAxis</label></td>
 						<td>
-							<input id="has-locked-axis" type="checkbox" bind:checked={$props.hasLockedAxis} />
+							<input id="has-locked-axis" type="checkbox" bind:checked={$rootProps.hasLockedAxis} />
 						</td>
 						<td></td>
 					</tr>
@@ -98,7 +100,7 @@
 						<td></td>
 						<td><label for="has-boundaries">hasBoundaries</label></td>
 						<td>
-							<input id="has-boundaries" type="checkbox" bind:checked={$props.hasBoundaries} />
+							<input id="has-boundaries" type="checkbox" bind:checked={$rootProps.hasBoundaries} />
 						</td>
 						<td></td>
 					</tr>
@@ -109,7 +111,7 @@
 							<input
 								id="can-clear-target-on-drag-out"
 								type="checkbox"
-								bind:checked={$props.canClearTargetOnDragOut}
+								bind:checked={$rootProps.canClearTargetOnDragOut}
 							/>
 						</td>
 						<td></td>
@@ -121,7 +123,7 @@
 							<input
 								id="can-remove-item-on-drop-out"
 								type="checkbox"
-								bind:checked={$props.canRemoveItemOnDropOut}
+								bind:checked={$rootProps.canRemoveItemOnDropOut}
 							/>
 						</td>
 						<td></td>
@@ -130,7 +132,7 @@
 						<td></td>
 						<td><label for="is-locked">isLocked</label></td>
 						<td>
-							<input id="is-locked" type="checkbox" bind:checked={$props.isLocked} />
+							<input id="is-locked" type="checkbox" bind:checked={$rootProps.isLocked} />
 						</td>
 						<td></td>
 					</tr>
@@ -138,7 +140,7 @@
 						<td></td>
 						<td><label for="is-disabled">isDisabled</label></td>
 						<td>
-							<input id="is-disabled" type="checkbox" bind:checked={$props.isDisabled} />
+							<input id="is-disabled" type="checkbox" bind:checked={$rootProps.isDisabled} />
 						</td>
 						<td></td>
 					</tr>
@@ -172,14 +174,14 @@
 			<a
 				class="app-nav__link"
 				href="/"
-				aria-current={$page.url.pathname === '/' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/' ? 'page' : undefined}
 			>
 				Basic
 			</a>
 			<a
 				class="app-nav__link"
 				href="/unstyled"
-				aria-current={$page.url.pathname === '/unstyled' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/unstyled' ? 'page' : undefined}
 				data-sveltekit-reload
 			>
 				Unstyled
@@ -187,119 +189,119 @@
 			<a
 				class="app-nav__link"
 				href="/no-animations"
-				aria-current={$page.url.pathname === '/no-animations' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/no-animations' ? 'page' : undefined}
 			>
 				No animations
 			</a>
 			<a
 				class="app-nav__link"
 				href="/direction-horizontal"
-				aria-current={$page.url.pathname === '/direction-horizontal' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/direction-horizontal' ? 'page' : undefined}
 			>
 				Direction horizontal
 			</a>
 			<a
 				class="app-nav__link"
 				href="/varying-heights"
-				aria-current={$page.url.pathname === '/varying-heights' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/varying-heights' ? 'page' : undefined}
 			>
 				Varying heights
 			</a>
 			<a
 				class="app-nav__link"
 				href="/with-handle"
-				aria-current={$page.url.pathname === '/with-handle' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/with-handle' ? 'page' : undefined}
 			>
 				With handle
 			</a>
 			<a
 				class="app-nav__link"
 				href="/with-drop-marker"
-				aria-current={$page.url.pathname === '/with-drop-marker' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/with-drop-marker' ? 'page' : undefined}
 			>
 				With drop marker
 			</a>
 			<a
 				class="app-nav__link"
 				href="/with-boundaries"
-				aria-current={$page.url.pathname === '/with-boundaries' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/with-boundaries' ? 'page' : undefined}
 			>
 				With boundaries
 			</a>
 			<a
 				class="app-nav__link"
 				href="/with-locked-axis"
-				aria-current={$page.url.pathname === '/with-locked-axis' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/with-locked-axis' ? 'page' : undefined}
 			>
 				With locked axis
 			</a>
 			<a
 				class="app-nav__link"
 				href="/dynamic-items"
-				aria-current={$page.url.pathname === '/dynamic-items' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/dynamic-items' ? 'page' : undefined}
 			>
 				Dynamic items
 			</a>
 			<a
 				class="app-nav__link"
 				href="/interactive-items"
-				aria-current={$page.url.pathname === '/interactive-items' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/interactive-items' ? 'page' : undefined}
 			>
 				Interactive items
 			</a>
 			<a
 				class="app-nav__link"
 				href="/locked-list"
-				aria-current={$page.url.pathname === '/locked-list' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/locked-list' ? 'page' : undefined}
 			>
 				Locked list
 			</a>
 			<a
 				class="app-nav__link"
 				href="/locked-items"
-				aria-current={$page.url.pathname === '/locked-items' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/locked-items' ? 'page' : undefined}
 			>
 				Locked items
 			</a>
 			<a
 				class="app-nav__link"
 				href="/disabled-list"
-				aria-current={$page.url.pathname === '/disabled-list' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/disabled-list' ? 'page' : undefined}
 			>
 				Disabled list
 			</a>
 			<a
 				class="app-nav__link"
 				href="/disabled-items"
-				aria-current={$page.url.pathname === '/disabled-items' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/disabled-items' ? 'page' : undefined}
 			>
 				Disabled items
 			</a>
 			<a
 				class="app-nav__link"
 				href="/inside-dialog"
-				aria-current={$page.url.pathname === '/inside-dialog' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/inside-dialog' ? 'page' : undefined}
 			>
 				Inside dialog
 			</a>
 			<a
 				class="app-nav__link"
 				href="/inside-custom-dialog"
-				aria-current={$page.url.pathname === '/inside-custom-dialog' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/inside-custom-dialog' ? 'page' : undefined}
 			>
 				Inside custom dialog
 			</a>
 			<a
 				class="app-nav__link"
 				href="/clear-target-on-drag-out"
-				aria-current={$page.url.pathname === '/clear-target-on-drag-out' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/clear-target-on-drag-out' ? 'page' : undefined}
 			>
 				Clear target on drag out
 			</a>
 			<a
 				class="app-nav__link"
 				href="/remove-item-on-drop-out"
-				aria-current={$page.url.pathname === '/remove-item-on-drop-out' ? 'page' : undefined}
+				aria-current={page.url.pathname === '/remove-item-on-drop-out' ? 'page' : undefined}
 			>
 				Remove item on drop out
 			</a>